sync: only close tasks with a ticket in resolved state
And if a ticket has only been Resolved for a little while, it might be
better to somehow hide its task, rather than completing it--because if it
gets unresolved, then the next sync will create a new task rather than
re-using the old one.

Actually, let's double-check the behaviour of a re-opened ticket; if we
can make it un-complete the task, we can revert this commit.